move implementation of the CLI utility to tabulate/cli.py
