+++ pyproject.toml
@@ -30,7 +30,7 @@ Homepage = "https://github.com/astanin/python-tabulate"
 widechars = ["wcwidth>=0.6.0"]
 
 [project.scripts]
-tabulate = "tabulate:_main"
+tabulate = "tabulate.cli:_main"
 
 [tool.flit.sdist]
 include = ["CHANGELOG", "test/", "tox.ini"]
