+++ tabulate/__main__.py
@@ -1,0 +1,3 @@
+from tabulate.cli import _main
+
+_main()
